+++ vortex-array/src/array/display/tree.rs
@@ -4,7 +4,6 @@
 use std::fmt::{self};
 
 use humansize::{DECIMAL, format_size};
-
 use crate::arrays::ChunkedEncoding;
 use crate::display::DisplayOptions;
 use crate::{Array, ArrayRef, ArrayVisitor};
@@ -25,7 +24,7 @@ impl fmt::Display for TreeDisplayWrapper {
 pub struct TreeFormatter<'a, 'b: 'a> {
     fmt: &'a mut fmt::Formatter<'b>,
     indent: String,
-    total_size: Option<usize>,
+    total_size: Option<u64>,
 }
 
 impl<'a, 'b: 'a> TreeFormatter<'a, 'b> {
